mesh-visualizer/requirements.txt
appdirs==1.4.4
argcomplete==2.0.0
bcrypt==4.0.1
Beaker==1.12.1
beautifulsoup4==4.12.3
blivet==3.8.2
blivet-gui==2.5.0
boto3==1.34.47
botocore==1.34.47
Brlapi==0.8.5
cachetools==5.3.1
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.3
cryptography==41.0.7
cson==0.8
cssselect==1.1.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distro==1.8.0
dnf==4.19.0
dropbox==11.36.2
duplicity==2.1.5
evdev==1.6.1
fasteners==0.19
fb-re2==1.0.7
fedora-third-party==0.10
file-magic==0.4.0
fros==1.1
future==0.18.3
git-filter-repo==2.38.0
google-api-core==2.11.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.60.0
gpg==1.20.0
grpcio==1.48.4
grpcio-status==1.48.4
httplib2==0.21.0
humanize==3.13.1
idna==3.4
inkex==1.3.0
jmespath==1.0.1
langtable==0.0.65
libcomps==0.1.20
libdnf==0.73.0
lxml==4.9.3
Mako==1.2.3
MarkupSafe==2.1.3
mercurial==6.5.3
nftables==0.1
numpy==1.24.4
oauth2client==4.1.3
olefile==0.46
packaging==23.1
paramiko==3.4.0
Paste==3.5.3
pexpect==4.8.0
pid==2.2.3
pillow==10.2.0
pipx==1.4.3
platformdirs==3.9.1
ply==3.11
podman-compose==1.0.6
productmd==1.38
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pwquality==1.4.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.2
PyDrive2==1.19.0
pyenchant==3.2.2
Pygments==2.15.1
PyGObject==3.46.0
pykickstart==3.48
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyparted==3.13.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyserial==3.5
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-meh==0.51
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
regex==2023.12.25
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
rpm==4.19.1.1
rpmconf==1.1.10
rsa==4.9
s-tui==1.1.6
s3transfer==0.10.0
scour==0.38.2
selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
setools==4.4.3
setuptools==67.7.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.6.0
soupsieve==2.5
speg==0.3
stone==3.2.1
systemd-python==235
Tempita==0.5.2
uritemplate==4.1.1
urllib3==1.26.18
urwid==2.1.2
userpath==1.9.1
wxPython==4.2.1
zombie-imp==0.0.2
